serverless/aws/routes/requirements.txt
html2text==2020.1.16
numpy==1.21.0
psycopg2-binary==2.8.6
pydantic==1.8.2
pinecone-client==2.0.3
requests==2.27.1
boto3==1.21.3
